move principal/authz annotations to core
