+++ backend/tissue-adapter/src/main/java/com/tissue/feature/activitylog/web/ActivityLogQueryController.java
@@ -8,8 +8,8 @@ import com.tissue.feature.sprint.domain.exception.SprintErrorCode;
 import com.tissue.global.openapi.IssueErrors;
 import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.SprintErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.IssueIdentifier;
 import com.tissue.shared.dto.KeysetPageResponse;
 import io.swagger.v3.oas.annotations.Operation;
