+++ backend/tissue-adapter/src/main/java/com/tissue/feature/comment/web/CommentCommandController.java
@@ -10,8 +10,8 @@ import com.tissue.feature.project.domain.exception.ProjectErrorCode;
 import com.tissue.global.openapi.CommentErrors;
 import com.tissue.global.openapi.IssueErrors;
 import com.tissue.global.openapi.ProjectErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.IssueIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
