+++ backend/tissue-adapter/src/main/java/com/tissue/feature/issue/web/IssueCommandController.java
@@ -33,8 +33,8 @@ import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.SprintErrors;
 import com.tissue.global.openapi.TagErrors;
 import com.tissue.global.openapi.WorkflowErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.BatchOperationResponse;
 import com.tissue.shared.dto.IssueIdentifier;
 import com.tissue.shared.dto.ProjectIdentifier;
