+++ backend/tissue-adapter/src/main/java/com/tissue/feature/issue/web/IssueQueryController.java
@@ -17,8 +17,8 @@ import com.tissue.feature.issue.web.request.IssueSearchRequest;
 import com.tissue.feature.project.domain.exception.ProjectErrorCode;
 import com.tissue.global.openapi.IssueErrors;
 import com.tissue.global.openapi.ProjectErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.CursorPage;
 import com.tissue.shared.dto.IssueIdentifier;
 import com.tissue.shared.dto.ProjectIdentifier;
