+++ backend/tissue-adapter/src/main/java/com/tissue/feature/issuetype/web/IssueFieldController.java
@@ -10,9 +10,9 @@ import com.tissue.feature.issuetype.web.request.UpdateIssueFieldRequest;
 import com.tissue.feature.member.domain.exception.MemberErrorCode;
 import com.tissue.global.openapi.IssueTypeErrors;
 import com.tissue.global.openapi.MemberErrors;
-import com.tissue.security.adapter.web.annotation.RequireSystemAdmin;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
+import com.tissue.shared.auth.RequireSystemAdmin;
 import com.tissue.shared.vo.Name;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
