+++ backend/tissue-adapter/src/main/java/com/tissue/feature/issuetype/web/IssueTypeController.java
@@ -11,9 +11,9 @@ import com.tissue.feature.workflow.domain.exception.WorkflowErrorCode;
 import com.tissue.global.openapi.IssueTypeErrors;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.WorkflowErrors;
-import com.tissue.security.adapter.web.annotation.RequireSystemAdmin;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
+import com.tissue.shared.auth.RequireSystemAdmin;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
