+++ backend/tissue-adapter/src/main/java/com/tissue/feature/issuetype/web/IssueTypeQueryController.java
@@ -5,8 +5,8 @@ import com.tissue.feature.issuetype.application.dto.response.IssueTypeSummary;
 import com.tissue.feature.issuetype.application.port.usecase.IssueTypeQueryUseCase;
 import com.tissue.feature.issuetype.domain.exception.IssueTypeErrorCode;
 import com.tissue.global.openapi.IssueTypeErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
