+++ backend/tissue-adapter/src/main/java/com/tissue/feature/member/web/MemberAdministrationController.java
@@ -6,9 +6,9 @@ import com.tissue.feature.member.web.request.AssignMemberTeamRequest;
 import com.tissue.feature.organization.team.domain.exception.TeamErrorCode;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.TeamErrors;
-import com.tissue.security.adapter.web.annotation.RequireSystemAdmin;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
+import com.tissue.shared.auth.RequireSystemAdmin;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
