+++ backend/tissue-adapter/src/main/java/com/tissue/feature/member/web/MemberProfileCommandController.java
@@ -8,8 +8,8 @@ import com.tissue.feature.member.web.request.UpdateMemberPositionRequest;
 import com.tissue.feature.organization.position.domain.exception.PositionErrorCode;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.PositionErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
