+++ backend/tissue-adapter/src/main/java/com/tissue/feature/member/web/MemberProfileQueryController.java
@@ -4,8 +4,8 @@ import com.tissue.feature.member.application.dto.MemberProfile;
 import com.tissue.feature.member.application.port.usecase.MemberProfileQueryUseCase;
 import com.tissue.feature.member.domain.exception.MemberErrorCode;
 import com.tissue.global.openapi.MemberErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
