+++ backend/tissue-adapter/src/main/java/com/tissue/feature/notification/web/NotificationCommandController.java
@@ -3,8 +3,8 @@ package com.tissue.feature.notification.web;
 import com.tissue.feature.notification.application.port.usecase.NotificationCommandUseCase;
 import com.tissue.feature.notification.domain.exception.NotificationErrorCode;
 import com.tissue.global.openapi.NotificationErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
