+++ backend/tissue-adapter/src/main/java/com/tissue/feature/notification/web/NotificationPreferenceController.java
@@ -3,8 +3,8 @@ package com.tissue.feature.notification.web;
 import com.tissue.feature.notification.application.dto.response.NotificationPreferenceResponse;
 import com.tissue.feature.notification.application.service.NotificationPreferenceService;
 import com.tissue.feature.notification.web.request.UpdateNotificationPreferenceRequest;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
