+++ backend/tissue-adapter/src/main/java/com/tissue/feature/notification/web/NotificationQueryController.java
@@ -2,8 +2,8 @@ package com.tissue.feature.notification.web;
 
 import com.tissue.feature.notification.application.dto.response.NotificationResponse;
 import com.tissue.feature.notification.application.port.usecase.NotificationQueryUseCase;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.KeysetPageResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
