+++ backend/tissue-adapter/src/main/java/com/tissue/feature/organization/team/web/TeamController.java
@@ -8,9 +8,9 @@ import com.tissue.feature.organization.team.web.request.CreateTeamRequest;
 import com.tissue.feature.organization.team.web.request.UpdateTeamRequest;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.TeamErrors;
-import com.tissue.security.adapter.web.annotation.RequireSystemAdmin;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
+import com.tissue.shared.auth.RequireSystemAdmin;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
