+++ backend/tissue-adapter/src/main/java/com/tissue/feature/project/web/ProjectController.java
@@ -8,8 +8,8 @@ import com.tissue.feature.project.web.request.CreateProjectRequest;
 import com.tissue.feature.project.web.request.UpdateProjectRequest;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.ProjectErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.ProjectIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
