+++ backend/tissue-adapter/src/main/java/com/tissue/feature/project/web/ProjectMemberController.java
@@ -7,8 +7,8 @@ import com.tissue.feature.project.domain.exception.ProjectErrorCode;
 import com.tissue.feature.project.web.request.AddProjectMembersRequest;
 import com.tissue.feature.project.web.request.ChangeRoleRequest;
 import com.tissue.global.openapi.ProjectErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.ProjectIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
