+++ backend/tissue-adapter/src/main/java/com/tissue/feature/sprint/web/SprintCommandController.java
@@ -12,8 +12,8 @@ import com.tissue.feature.sprint.web.request.StartSprintRequest;
 import com.tissue.feature.sprint.web.request.UpdateSprintRequest;
 import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.SprintErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.ProjectIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
