+++ backend/tissue-adapter/src/main/java/com/tissue/feature/sprint/web/SprintQueryController.java
@@ -9,8 +9,8 @@ import com.tissue.feature.sprint.domain.SprintStatus;
 import com.tissue.feature.sprint.domain.exception.SprintErrorCode;
 import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.SprintErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.ProjectIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
