+++ backend/tissue-adapter/src/main/java/com/tissue/feature/tag/web/TagCommandController.java
@@ -8,8 +8,8 @@ import com.tissue.feature.tag.web.request.CreateTagRequest;
 import com.tissue.feature.tag.web.request.UpdateTagRequest;
 import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.TagErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.ProjectIdentifier;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
