+++ backend/tissue-adapter/src/main/java/com/tissue/feature/vcs/web/github/GithubIntegrationController.java
@@ -9,8 +9,8 @@ import com.tissue.feature.vcs.domain.enums.VcsProvider;
 import com.tissue.feature.vcs.domain.exception.VcsErrorCode;
 import com.tissue.global.openapi.ProjectErrors;
 import com.tissue.global.openapi.VcsErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
