+++ backend/tissue-adapter/src/main/java/com/tissue/feature/wiki/web/WikiAttachmentController.java
@@ -6,8 +6,8 @@ import com.tissue.feature.wiki.application.dto.response.WikiAttachmentUploadResp
 import com.tissue.feature.wiki.application.port.usecase.WikiAttachmentUseCase;
 import com.tissue.feature.wiki.domain.exception.WikiErrorCode;
 import com.tissue.global.openapi.WikiErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
