+++ backend/tissue-adapter/src/main/java/com/tissue/feature/wiki/web/WikiBookmarkCommandController.java
@@ -3,8 +3,8 @@ package com.tissue.feature.wiki.web;
 import com.tissue.feature.wiki.application.port.usecase.WikiBookmarkCommandUseCase;
 import com.tissue.feature.wiki.domain.exception.WikiErrorCode;
 import com.tissue.global.openapi.WikiErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
