+++ backend/tissue-adapter/src/main/java/com/tissue/feature/wiki/web/WikiDocumentCommandController.java
@@ -9,8 +9,8 @@ import com.tissue.feature.wiki.web.request.SetDocumentParentRequest;
 import com.tissue.feature.wiki.web.request.UpdateDocumentContentRequest;
 import com.tissue.feature.wiki.web.request.UpdateDocumentTitleRequest;
 import com.tissue.global.openapi.WikiErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
