+++ backend/tissue-adapter/src/main/java/com/tissue/feature/wiki/web/WikiDocumentQueryController.java
@@ -9,8 +9,8 @@ import com.tissue.feature.wiki.application.dto.response.WikiSnapshotSummary;
 import com.tissue.feature.wiki.application.port.usecase.WikiQueryUseCase;
 import com.tissue.feature.wiki.domain.exception.WikiErrorCode;
 import com.tissue.global.openapi.WikiErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.dto.KeysetPageResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
