+++ backend/tissue-adapter/src/main/java/com/tissue/feature/workflow/web/WorkflowCommandController.java
@@ -14,9 +14,9 @@ import com.tissue.feature.workflow.web.request.UpdateWorkflowRequest;
 import com.tissue.feature.workflow.web.request.UpdateWorkflowVcsSettingsRequest;
 import com.tissue.global.openapi.MemberErrors;
 import com.tissue.global.openapi.WorkflowErrors;
-import com.tissue.security.adapter.web.annotation.RequireSystemAdmin;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
+import com.tissue.shared.auth.RequireSystemAdmin;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
