+++ backend/tissue-adapter/src/main/java/com/tissue/feature/workflow/web/WorkflowQueryController.java
@@ -6,8 +6,8 @@ import com.tissue.feature.workflow.application.dto.response.WorkflowSummary;
 import com.tissue.feature.workflow.application.port.usecase.WorkflowQueryUseCase;
 import com.tissue.feature.workflow.domain.exception.WorkflowErrorCode;
 import com.tissue.global.openapi.WorkflowErrors;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
