+++ backend/tissue-bootstrap/src/test/java/com/tissue/feature/wiki/WikiCommandServiceIntegrationTest.java
@@ -15,7 +15,7 @@ import com.tissue.feature.wiki.application.service.WikiCommandService;
 import com.tissue.feature.wiki.domain.WikiDocument;
 import com.tissue.feature.wiki.domain.WikiDocumentSnapshot;
 import com.tissue.feature.wiki.domain.enums.SemanticUpdateType;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.shared.exception.base.BadRequestException;
 import com.tissue.shared.exception.base.ForbiddenException;
 import com.tissue.support.IntegrationTestSupport;
