+++ backend/tissue-bootstrap/src/test/java/com/tissue/feature/workflow/WorkflowCommandControllerSecurityTest.java
@@ -4,7 +4,7 @@ import static org.springframework.security.test.web.servlet.request.SecurityMock
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.MemberDetails;
 import com.tissue.support.IntegrationTestSupport;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
