+++ backend/tissue-security/src/main/java/com/tissue/security/adapter/web/AuthenticationController.java
@@ -10,8 +10,8 @@ import com.tissue.security.application.dto.response.LoginResponse;
 import com.tissue.security.application.dto.response.RefreshTokenResponse;
 import com.tissue.security.application.port.usecase.AuthenticationUseCase;
 import com.tissue.security.domain.exception.AuthenticationErrorCode;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
