+++ backend/tissue-security/src/main/java/com/tissue/security/adapter/web/MemberAccountController.java
@@ -13,8 +13,8 @@ import com.tissue.security.adapter.web.request.UpdateMemberUsernameRequest;
 import com.tissue.security.adapter.web.request.WithdrawMemberRequest;
 import com.tissue.security.application.port.usecase.MemberAccountUseCase;
 import com.tissue.security.domain.exception.AuthenticationErrorCode;
-import com.tissue.security.principal.CurrentMember;
-import com.tissue.security.principal.MemberDetails;
+import com.tissue.shared.auth.CurrentMember;
+import com.tissue.shared.auth.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
