+++ backend/tissue-security/src/main/java/com/tissue/security/application/service/AuthenticationService.java
@@ -10,8 +10,8 @@ import com.tissue.security.application.port.usecase.AuthenticationUseCase;
 import com.tissue.security.domain.TokenProvider;
 import com.tissue.security.domain.exception.RefreshTokenNotFoundException;
 import com.tissue.security.domain.exception.TokenReuseDetectedException;
-import com.tissue.security.principal.MemberDetails;
 import com.tissue.security.util.TokenHashUtil;
+import com.tissue.shared.auth.MemberDetails;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
