+++ backend/tissue-security/src/test/java/com/tissue/security/AuthenticationServiceTest.java
@@ -20,8 +20,8 @@ import com.tissue.security.application.service.TokenPairCreateService;
 import com.tissue.security.domain.TokenProvider;
 import com.tissue.security.domain.exception.RefreshTokenNotFoundException;
 import com.tissue.security.domain.exception.TokenReuseDetectedException;
-import com.tissue.security.principal.MemberDetails;
 import com.tissue.security.util.TokenHashUtil;
+import com.tissue.shared.auth.MemberDetails;
 import java.util.Collections;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
